+++ .cli/.preview/.demo/.templates/.njk/templates/modules/m04-manual/index.ts
@@ -5,7 +5,7 @@
  *
  */
 
-import { IModuleInstance } from '@scripts/models/models.d.ts';
+import { IModuleInstance } from '@scripts/models/models.d';
 import createModule from '@scripts/libs/create-module';
 // import pubsub from '@scripts/helpers/pubsub';
 import helperDom from '@scripts/helpers/helper-dom';