+++ config/.darvinconf.dev.js
@@ -1,4 +1,4 @@
-global.project = 'Darvin';
+global.project = 'Darvin Webpack Boilerplate';
 global.mainChunk = 'scripts/main';
 global.proxy = '';
 global.port = '8001';