+++ webpack/settings/env-devserver/index.js
@@ -5,18 +5,42 @@ let devServer = require('../../libs/devserver-storage');
 
 const dev = {
   devServer: {
-    before(app, server) {
-      // set storage
-      devServer.server = server;
-      devServer.app = app;
+    devMiddleware: {
+      writeToDisk: true
     },
-    contentBase: path.resolve(basePath, 'dist'),
-    watchContentBase: true,
+    onBeforeSetupMiddleware: function (devServer) {
+      if (!devServer) {
+        throw new Error('webpack-dev-server is not defined');
+      }
+
+      // set storage
+      devServer.server = devServer;
+      devServer.app = devServer.app;
+    },
+    static: [
+      {
+        directory: path.join(basePath, '/dist/assets'),
+        publicPath: '/assets',
+      },
+      {
+        directory: path.join(basePath, '/dist/preview'),
+        publicPath: '/preview',
+      },
+    ],
+    client: {
+      logging: 'info',
+      progress: false,
+      overlay: {
+        errors: true,
+        warnings: false,
+      },
+    },
+    https: false,
     compress: true,
     port: global.port,
     open: true,
-    writeToDisk: true,
-    index: 'index.html',
+    hot: true,
+    liveReload: true,
     headers: {
       'Access-Control-Allow-Origin': '*',
       'Access-Control-Allow-Headers': '*',
@@ -26,6 +50,14 @@ const dev = {
         target: "http://localhost:8002",
         secure: false
       }
+    },
+    onListening: function (devServer) {
+      if (!devServer) {
+        throw new Error('webpack-dev-server is not defined');
+      }
+
+      const port = devServer.server.address().port;
+      console.log('|> Darvin running on port:', port);
     },
   }
 }