+++ webpack/settings/javascript-copy-vendors/index.js
@@ -2,25 +2,29 @@ const CopyWebpackPlugin = require('copy-webpack-plugin');
 const basePath = process.cwd();
 
 const copyConfig = {
-  from: basePath + '/' + global.inputDirs.src + '/scripts/vendors/*.js',
-  to: global.server.assets + '/scripts/vendors/',
-  flatten: true,
-  cache: true
+  from: `${basePath}/${global.inputDirs.src}/scripts/vendors/*.js`,
+  to() {
+    return `${global.server.assets}/scripts/vendors/[name][ext]`;
+  },
 };
 
 const prod = {
   plugins: [
-    new CopyWebpackPlugin([
-      copyConfig,
-    ], {})
+    new CopyWebpackPlugin({
+      patterns: [
+        copyConfig
+      ]
+    })
   ],
 };
 
 const dev = {
   plugins: [
-    new CopyWebpackPlugin([
-      copyConfig,
-    ], {})
+    new CopyWebpackPlugin({
+      patterns: [
+        copyConfig
+      ]
+    })
   ],
 };