+++ webpack/settings/javascript/babel-config.js
@@ -6,99 +6,35 @@ const CACHE_PATH = path.join(ROOT_PATH, 'tmp/cache');
 const modern = {
   presets: [
     ["@babel/preset-env", {
-      targets: {
-        browsers: [
-          "> 1% in CH",
-          "not ie 11",
-          "not dead"
-        ]
-      },
       useBuiltIns: "usage",
-      corejs: 3,
-    }]
+      corejs: 3
+    }],
   ],
   plugins: [
-    "@babel/plugin-syntax-dynamic-import",
-    "transform-eval"
+    "@babel/plugin-proposal-class-properties"
   ],
-  comments: false,
+  comments: true,
   cacheDirectory: path.join(CACHE_PATH, 'babel-loader')
 };
-
-const legacy = {
-  presets: [
-    ["@babel/preset-env", {
-      targets: {
-        browsers: [
-          "> 1% in CH",
-          "ie >= 11",
-          "not dead"
-        ]
-      },
-      useBuiltIns: "usage",
-      corejs: 3,
-    }]
-  ],
-  plugins: [
-    "@babel/plugin-syntax-dynamic-import",
-    "transform-eval"
-  ],
-  comments: false,
-  cacheDirectory: path.join(CACHE_PATH, 'babel-loader')
-};
-
 
 const modernReact = {
   presets: [
-    ["@babel/preset-env", {
-      targets: {
-        browsers: [
-          "> 1% in CH",
-          "not ie 11",
-          "not dead"
-        ]
-      },
-      useBuiltIns: "usage",
-      corejs: 3,
+    ["@babel/preset-react", {
+      runtime: "automatic"
     }],
-    "@babel/preset-react"
+    ["@babel/preset-env", {
+      useBuiltIns: "usage",
+      corejs: 3
+    }],
   ],
   plugins: [
-    "@babel/plugin-syntax-dynamic-import",
-    "@babel/plugin-proposal-class-properties",
-    "transform-eval"
+    "@babel/plugin-proposal-class-properties"
   ],
-  comments: false,
-  cacheDirectory: path.join(CACHE_PATH, 'babel-loader')
-};
-
-const legacyReact = {
-  presets: [
-    ["@babel/preset-env", {
-      targets: {
-        browsers: [
-          "> 1% in CH",
-          "ie >= 11",
-          "not dead"
-        ]
-      },
-      useBuiltIns: "usage",
-      corejs: 3,
-    }],
-    "@babel/preset-react"
-  ],
-  plugins: [
-    "@babel/plugin-syntax-dynamic-import",
-    "@babel/plugin-proposal-class-properties",
-    "transform-eval"
-  ],
-  comments: false,
+  comments: true,
   cacheDirectory: path.join(CACHE_PATH, 'babel-loader')
 };
 
 module.exports = {
   modern,
-  legacy,
-  modernReact,
-  legacyReact
+  modernReact
 };