+++ changelog.md
@@ -1,5 +1,9 @@
 # Release Changelog
 
+## v3.0.0
+- upgrade settings to webpack 5
+- remove legacy scripts
+
 ## v2.0.2
 - clean copy assets for proxy
 - add picturefill to default package